pyproject.toml
[project]
name = "eq-test-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "KhangPTT373",email = "phmtrntuankhang@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "openai (>=1.98.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "authlib (>=1.6.1,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.17.1,<2.0.0)",
    "bandit (>=1.8.6,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "prometheus-client (>=0.22.1,<0.23.0)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
    "redis (>=6.4.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".mypy_cache",
    ".pytest_cache"
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101", "B601"]

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.4"

